requirements.txt
streamlit
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
google-generativeai